+++ xugrid/data/__init__.py
@@ -1,2 +1,2 @@
-from .sample_data import adh_san_diego, elevation_nl, xoxo
-from .synthetic import disk, generate_disk
+from xugrid.data.sample_data import adh_san_diego, elevation_nl, xoxo
+from xugrid.data.synthetic import disk, generate_disk
