+++ xugrid/ugrid/align.py
@@ -7,6 +7,7 @@ Align UGRID topology with the xarray object and vice versa.
 
 """
 
+
 def ugrid_aligns(obj, grid):
     """
     Check whether the xarray object dimensions still align with the grid.
